Refactor: Group specialized agent modules under agent.agents
- Moved ErrorAnalysisAgent, ErrorCorrectionAgent, and PerformanceAnalysisAgent into the new `agent/agents/` package.
- Extracted ArchitectAgent and MaestroAgent from the old `agent/agents.py` into their own files within `agent/agents/`.
- Moved the shared `parse_json_response` function from the old `agent/agents.py` to `agent/utils/json_parser.py`.
- Updated `agent/agents/__init__.py` to export all relevant agent classes.
- Updated import statements and `@patch` decorator paths across the codebase (including relevant test files) to reflect the new locations.
- Deleted the original `agent/agents.py` file as it became redundant.

Note: Pytest execution was hindered by a persistent sandbox environment issue, but manual code and structure verification indicates changes are correctly applied.
