+++ agent/agents/__init__.py
@@ -1,0 +1,15 @@
+# This file makes the 'agents' directory a Python package.
+
+from .architect_agent import ArchitectAgent
+from .maestro_agent import MaestroAgent
+from .error_analyzer import ErrorAnalysisAgent
+from .error_correction import ErrorCorrectionAgent
+from .performance_analyzer import PerformanceAnalysisAgent
+
+__all__ = [
+    "ArchitectAgent",
+    "MaestroAgent",
+    "ErrorAnalysisAgent",
+    "ErrorCorrectionAgent",
+    "PerformanceAnalysisAgent",
+]
