--- example_config.json
@@ -1,10 +1,0 @@
-{
-  "api_key": "your-api-key-here",
-  "default_model": "gpt-3.5-turbo",
-  "temperature": 0.7,
-  "base_url": "https://openrouter.ai/api/v1",
-  "log_level": "INFO",
-  "max_retries": 3,
-  "timeout_seconds": 30,
-  "memory_file": "HEPHAESTUS_MEMORY.json"
-}
